Move information_schema DatabaseSchema implementation to own file
